캐시된 이미지 파일의 갯수를 리턴할 때 size_t 로 하도록 변경.

+++ ZViewer/ZViewer/src/ZCacheImage.h
@@ -29,7 +29,7 @@ public:
 
 	void StartThread();
 
-	int GetNumOfCacheImage() const { return m_cacheData.size(); 	}
+	size_t GetNumOfCacheImage() const { return m_cacheData.size(); 	}
 
 	static DWORD WINAPI ThreadFuncProxy(LPVOID p);
 	void ThreadFunc();
